pedestrain/bounding_box_prediction/requirements.txt
#torch==1.4.0
#torchvision==0.5.0
numpy==1.16.1
matplotlib
pillow
jupyter
pandas
# sklearn
scikit-learn 
scipy
opencv-python